refactor: 优化 CMake 构建系统并修复多处缓存管理问题
主要变更：
- 将自定义 find_package 模块替换为 pkg-config 查找依赖
- 优化 buffstruct 类实现，提高内存管理效率
- 增加下载线程数量 (5->20) 提升并行性能
- 修复内存泄漏和数据竞争问题
- 修复断言条件错误和类型不匹配问题
- 添加文件系统状态缓存以减少重复调用

修复了多处潜在的内存安全问题，并提高了文件下载和缓存管理的性能。

summarized by claude sonnet 3.7

+++ include/fmdisk.h
@@ -14,7 +14,7 @@ int fm_prepare();
 
 int fm_statfs(struct statvfs* sf);
 
-int fm_download(const filekey& file, size_t startp, size_t len, buffstruct& bs);
+int fm_download(const filekey& file, off_t startp, size_t len, buffstruct& bs);
 
 //upload `file` in `fileat`, return key in `file`
 int fm_upload(const filekey& fileat, filekey& file, const char* data, size_t len, bool overwrite);
